# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Stellar Trust Escrow API
  version: 0.1.0
  description: |
    REST API for the Stellar Trust Escrow platform — a trust-minimised escrow
    service built on the Stellar blockchain.

    ## Authentication
    Most endpoints require a Bearer JWT token in the `Authorization` header.
    Admin endpoints additionally require the `x-admin-api-key` header.

    ## Pagination
    All collection endpoints use **cursor-based pagination** and return:
    ```
    {
      "data": [...],
      "next_cursor": "<opaque-base64url-token>",
      "has_more": true
    }
    ```
    Pass `?cursor=<next_cursor>` to fetch the next page. The cursor encodes the
    sort key and direction and is stable across concurrent inserts.

    ## Idempotency Keys
    All `POST` and `PATCH` endpoints accept an optional `Idempotency-Key` header.
    Supplying this header makes the call safe to retry:
    - The response is cached in Redis for **24 hours** under `(tenantId, method, path, key)`.
    - Duplicate requests within the TTL return the cached response immediately
      (status `X-Idempotency-Replayed: true`).
    - If a duplicate arrives while the first is still in-flight, a **409 Conflict** is returned.
    - Use any unique string (UUID recommended): `Idempotency-Key: <uuid-v4>`

    ## Environments
    - **Testnet**: `https://api.test.stellar-trust-escrow.example.com`
    - **Mainnet**: `https://api.stellar-trust-escrow.example.com`
  contact:
...
  license:
    name: AGPL-3.0
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /api/notifications/events:
    post:
      tags: [Notifications]
      summary: Enqueue an email notification event
      operationId: enqueueNotification
      security:
        - AdminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [eventType, data]
              properties:
                eventType:
                  type: string
                  enum:
                    - escrow.status_changed
                    - milestone.completed
                    - dispute.raised
                  example: milestone.completed
                data:
                  type: object
                  required: [recipients]
                  properties:
                    recipients:
                      type: array
                      items:
                        type: string
                        format: email
                      example: ['user@example.com']
                    escrowId:
                      type: string
                    milestoneId:
                      type: string
      responses:
        '202':
          description: Event accepted for delivery
          content:
            application/json:
              schema:
                type: object
                properties:
                  enqueued:
                    type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/notifications/subscribe:
    post:
      tags: [Notifications]
      summary: Re-subscribe an email to notifications using the unsubscribe token
      operationId: subscribe
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, token]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                token:
                  type: string
                  description: Unsubscribe token from the original email link
      responses:
        '200':
          description: Resubscribed
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                    format: email
                  unsubscribedAt:
                    type: string
                    format: date-time
                    nullable: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          description: Invalid resubscribe token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/notifications/unsubscribe:
    get:
      tags: [Notifications]
      summary: Unsubscribe via email link (GET for browser redirect)
      operationId: unsubscribeGet
      security: []
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
        - name: token
          in: query
          required: true
          schema:
            type: string
        - name: reason
          in: query
          schema:
            type: string
      responses:
        '200':
          description: HTML confirmation page
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Missing email or token
          content:
            text/html:
              schema:
                type: string
    post:
      tags: [Notifications]
      summary: Unsubscribe via API
      operationId: unsubscribePost
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, token]
              properties:
                email:
                  type: string
                  format: email
                token:
                  type: string
                reason:
                  type: string
      responses:
        '200':
          description: Unsubscribed
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                    format: email
                  unsubscribedAt:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/notifications/queue:
    get:
      tags: [Notifications]
      summary: Get email queue snapshot (size, processing stats)
      operationId: getQueueSnapshot
      security:
        - AdminAuth: []
      responses:
        '200':
          description: Queue snapshot
          content:
            application/json:
              schema:
                type: object
                properties:
                  pending:
                    type: integer
                  processing:
                    type: integer
                  processed:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/events:
    get:
      tags: [Events]
      summary: List all indexed on-chain events
      operationId: listEvents
      security: []
      parameters:
        - name: eventType
          in: query
          schema:
            type: string
          example: escrow_created
        - name: escrowId
          in: query
          schema:
            type: string
        - name: fromLedger
          in: query
          schema:
            type: integer
          description: Inclusive lower bound ledger sequence
        - name: toLedger
          in: query
          schema:
            type: integer
          description: Inclusive upper bound ledger sequence
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated event list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedEvents'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /api/auth/login
    AdminAuth:
      type: apiKey
      in: header
      name: x-admin-api-key
      description: Admin API key for protected admin endpoints
  parameters:
    Page:
      name: page
      in: query
      description: 1-based page number. Values below 1 are normalised to 1.
      schema:
        type: integer
        default: 1
        minimum: 1
    Limit:
      name: limit
      in: query
      description: Page size. Defaults to 20 and is capped at 100.
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Something went wrong
        errorId:
          type: string
          description: Sentry event ID for support correlation
    PaginatedResponse:
      type: object
      required: [data, page, limit, total, totalPages, hasNextPage, hasPreviousPage]
      properties:
        data:
          type: array
          items: {}
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 42
        totalPages:
          type: integer
          example: 3
        hasNextPage:
          type: boolean
          example: true
        hasPreviousPage:
          type: boolean
          example: false
...
    IndexedEvent:
      type: object
      properties:
        id:
          type: integer
          format: int64
        ledger:
          type: integer
        txHash:
          type: string
        eventType:
          type: string
        escrowId:
          type: string
          nullable: true
        data:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
    PaginatedEvents:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/IndexedEvent'
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...